[Test] Unify to the failing test with TARGET_BACKEND: JVM_IR
wrt https://github.com/JetBrains/kotlin/pull/5325#discussion_r1661531383
