+++ compiler/testData/codegen/box/reflection/callBy/inlineClasses/nullableObject/defaultArguments/jvmStaticFunctionsOnObject.kt
@@ -1,12 +1,12 @@
-// TARGET_BACKEND: JVM
+// TARGET_BACKEND: JVM_IR
 // JVM_TARGET: 1.8
-// IGNORE_BACKEND: JVM_IR
 // WITH_REFLECT
 
 import kotlin.reflect.KCallable
 import kotlin.reflect.KParameter
 import kotlin.reflect.full.instanceParameter
 import kotlin.test.assertEquals
+import kotlin.test.assertFailsWith
 
 @JvmInline
 value class S(val value: String?) {
@@ -100,7 +100,9 @@ fun box(): String {
     val cMembers = C::class.members.associateBy { it.name }
     assertEquals(S("124"), cMembers.getValue("foo").callBy(C, one, "2", four))
     assertEquals(four, cMembers.getValue("staticDefault1_1").callBy(C, four))
-    assertEquals(default, cMembers.getValue("staticDefault1_1").callByEmpty(C))
+    assertFailsWith<Error>("Remove assertFailsWith and try again, as this problem may have been fixed.") {
+        assertEquals(default, cMembers.getValue("staticDefault1_1").callByEmpty(C))
+    }
     assertEquals(four, cMembers.getValue("staticDefault1_2").callBy(C, four))
     assertEquals(default, cMembers.getValue("staticDefault1_2").callByEmpty(C))
     assertEquals(
@@ -113,7 +115,9 @@ fun box(): String {
             0L, zero
         )
     )
-    assertEquals(C.staticDefault32_1(), cMembers.getValue("staticDefault32_1").callByEmpty(C))
+    assertFailsWith<Error>("Remove assertFailsWith and try again, as this problem may have been fixed.") {
+        assertEquals(C.staticDefault32_1(), cMembers.getValue("staticDefault32_1").callByEmpty(C))
+    }
     assertEquals(
         S("00"),
         cMembers.getValue("staticDefault32_2").callBy(
@@ -135,7 +139,9 @@ fun box(): String {
             0L, 0L, zero
         )
     )
-    assertEquals(C.staticDefault33_1(), cMembers.getValue("staticDefault33_1").callByEmpty(C))
+    assertFailsWith<Error>("Remove assertFailsWith and try again, as this problem may have been fixed.") {
+        assertEquals(C.staticDefault33_1(), cMembers.getValue("staticDefault33_1").callByEmpty(C))
+    }
     assertEquals(
         S("00"),
         cMembers.getValue("staticDefault33_2").callBy(
@@ -150,7 +156,9 @@ fun box(): String {
 
     assertEquals(S("124"), C::foo.callBy(one, "2", four))
     assertEquals(four, C::staticDefault1_1.callBy(four))
-    assertEquals(default, C::staticDefault1_1.callBy(emptyMap()))
+    assertFailsWith<Error>("Remove assertFailsWith and try again, as this problem may have been fixed.") {
+        assertEquals(default, C::staticDefault1_1.callBy(emptyMap()))
+    }
     assertEquals(four, C::staticDefault1_2.callBy(four))
     assertEquals(default, C::staticDefault1_2.callBy(emptyMap()))
     assertEquals(
@@ -162,7 +170,9 @@ fun box(): String {
             0L, zero
         )
     )
-    assertEquals(C.staticDefault32_1(), C::staticDefault32_1.callBy(emptyMap()))
+    assertFailsWith<Error>("Remove assertFailsWith and try again, as this problem may have been fixed.") {
+        assertEquals(C.staticDefault32_1(), C::staticDefault32_1.callBy(emptyMap()))
+    }
     assertEquals(
         S("00"),
         C::staticDefault32_2.callBy(
@@ -182,7 +192,9 @@ fun box(): String {
             0L, 0L, zero
         )
     )
-    assertEquals(C.staticDefault33_1(), C::staticDefault33_1.callBy(emptyMap()))
+    assertFailsWith<Error>("Remove assertFailsWith and try again, as this problem may have been fixed.") {
+        assertEquals(C.staticDefault33_1(), C::staticDefault33_1.callBy(emptyMap()))
+    }
     assertEquals(
         S("00"),
         C::staticDefault33_2.callBy(
